rename .NET app to DeviceManager, delete IMessagePublisher
