+++ sample-application/device-manager/DeviceManager/DeviceManager.csproj
