+++ sample-application/device-manager/DeviceManager/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 {
   "profiles": {
-    "SensorDataGenerator": {
+    "DeviceManager": {
       "commandName": "Project"
     },
     "Docker": {
